Add BeanFactory context generator
This commit adds the necessary infrastructure to generate code that
represents the state of a particular BeanFactory.

`BeanValueWriter` provides an abstraction to write the instance supplier
of a particular bean definition. For corner cases, a
`BeanValueWriterSupplier` implementation can be registered in
`META-INF/spring.factories` to return a specific implementation.

The generated code relies on helper classes in the infrastructure
package. Those should be made available when compiling the generated
code and at runtime.